apps/api/requirements.txt

fastapi
uvicorn[standard]
pydantic
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
cryptography
pytest
pytest-mock
# Auth
pyjwt
passlib[bcrypt]
# DB
pymongo
pymysql
sqlalchemy-pytds
# AI
google-generativeai

# Oracle
oracledb

# ClickHouse
clickhouse-connect
clickhouse-sqlalchemy
clickhouse-driver

# Analytics (Offline-First)
duckdb
duckdb-engine

# Utils
redis
cloudinary
psutil
pandas
pyarrow
fastparquet
openpyxl
